Fix validation HTML output
The plots section of the HTML output for the validation tests were not showing up due to:

* `plots_section.html` expected a dictionary and got a list (behavior was changed when numerics
was integrated). So, I split this file into test specific versions as was done in performance.

* the test images paths weren't relative paths

Also, because systems are now starting to adopt python3 as default, I changed the hashbang to call
python2 explicitly.